Add dev-% and make Makefile.stage2 autonomous
Add the directive dev-% which is a set of convenience targets for
assembling the source directory with its debian directory added in
build/osrelease. For example, one would do:

    make dev-unstable

With this addition, one would expect to be able to use the makefile
under this directory to ease one's work. To accomplish this, we need to
remove the need for exported variables from the stage1 makefile.

In the stage1 makefile, export PROJECT, VERSION and SOURCE_V to the
disk. Those are exported in files _PROJECT_, _VERSION_ and _SOURCE_V_,
respectively, under the build/ directory. The stage2 makefile now has to
simply 'cat' their contents to set values for the right variables.

The DISTRO variable doesn't really need to receive information from
stage1 since the stage2 makefile is placed and run under a directory
named after the value that DISTRO receives. Get the information from the
directory name and be gone with the exported variable.

Also, to make the stage2 makefile more usable for a human, add 'deb',
'dsc' and 'clean' convenience targets. And, since we already have
information from the DISTRO variable, change 'upload-%' to 'upload',
which is easier to call by hand.