Add 'apply-%' in stage1 and 'apply' in stage2
The dev-% targets would be pretty useless if you had to bring back
changes manully. Add 'apply-%' in Makefile, that copies all
modifications that are not identical to files in common/ to the
*version*/ directory.

Also add a convenience 'apply' target in Makefile.stage2 that calls the
appropriate apply-% in Makefile.

Finally, add a note about this in the README and make sure to warn
people that its implementation is _dangerous_ for the files' contents.